+++ src/main/java/blackjack/dto/Participants.java
@@ -1,5 +1,11 @@
-package blackjack.domain;
+package blackjack.dto;
 
+import blackjack.domain.Cards;
+import blackjack.domain.Dealer;
+import blackjack.domain.Outcome;
+import blackjack.domain.Playable;
+import blackjack.domain.Player;
+import blackjack.domain.Players;
 import java.util.*;
 import java.util.stream.Collectors;
