+++ src/test/java/blackjack/DealerTest.java
@@ -1,16 +1,5 @@
 package blackjack;
 
-import blackjack.domain.Card;
-import blackjack.domain.Dealer;
-import blackjack.domain.Playable;
-import blackjack.utils.CardDeck;
-import blackjack.utils.FixedCardDeck;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import java.util.Arrays;
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class DealerTest {/*
